Seperated view page to different views.
