+++ src/views/view.js
@@ -1,0 +1,3 @@
+import renderInfo from './info';
+
+document.onload(renderInfo());
